4_langgraph/requirements_local_agent.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0 